Fix completion suggestions not updating as more is typed.
TextChangedP is run when texted is inserted but the preview window is open.

I don't know why, but it seems like this was designed to only send the
completion request one time after the initial trigger, even though we can get
better result by continuing to send new chars to the language server.

This check is also removed.

With these changes I get updates as I type and don't need to baskspace and
insert to get to updates.